LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/semantics/popovers/popover-not-keyboard-focusable-expected.txt

Button 1
This is a popover without a focusable element

 Button 2

FAIL Popover should not be keyboard focusable assert_equals: Keyboard focus should skip the open popover expected Element node <button id="secondfocus" tabindex="0">Button 2</button> but got Element node <button id="firstfocus" tabindex="0">Button 1</button>
